requirements.txt
# HTTP and scraping
requests>=2.31.0
requests-cache>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0

# AI search
anthropic>=0.18.0
